Migrate admin dashboard page to app router
