+++ package.json
@@ -47,7 +47,8 @@
     "semver": "^7.5.2",
     "word-wrap": "^1.2.4",
     "@babel/traverse": "7.23.2",
-    "parse-url": "8.1.0"
+    "parse-url": "8.1.0",
+    "domino": "Haringat/domino#2721294c75334ece635dfc72cad0a3dede7391fb"
   },
   "packageManager": "yarn@3.6.0"
 }
