+++ packages/admin/setupJestEnv.js
@@ -1,3 +1,4 @@
 process.env.SESSION_COOKIE_NAME = 'sessionId';
 process.env.JWT_COOKIE_NAME = 'user-service-token';
 process.env.SUB_PATH = '/apply/admin';
+process.env.ONE_LOGIN_MIGRATION_JOURNEY_ENABLED = true;
