+++ packages/gap-web-ui/package.json
@@ -61,6 +61,9 @@
     "typescript": "^4.7.4"
   },
   "dependencies": {
-    "turndown": "^7.1.1"
+    "turndown": "^7.1.2"
+  },
+  "resolutions": {
+    "domino": "Haringat/domino#2721294c75334ece635dfc72cad0a3dede7391fb"
   }
 }
